added switch between creating a Dockerfile or a Definition file depending on if the backend is singularity
