+++ e4s_alc/cli/commands/create.py
@@ -87,4 +87,3 @@ class CreateCommand(AbstractCommand):
         
         self.check_arguments(arg_namespace)
         model = CreateModel(arg_namespace)
-        model.create()
