+++ e4s_alc/model/__init__.py
@@ -1,1 +1,3 @@
 from e4s_alc.model.model import Model
+from e4s_alc.model.create_dockerfile import CreateDockerfileModel
+from e4s_alc.model.create_definitionfile import CreateDefinitionfileModel
